Restore PixiJS App and add logging for debugging
Restore the PixiJS version of App.tsx and add comprehensive console logging across the App and GameComponent to help track rendering, game state, start screen rendering, and PixiJS initialization. This includes reintroducing the start screen UI, player count handling, and wiring to the PixiGameRenderer, plus a small backup App-test to verify React rendering.

These changes were needed to isolate a blank-page issue to the PixiJS integration or Tailwind loading by providing detailed runtime logs and a minimal test harness for reproduction.
